requirements.txt
fastapi
uvicorn
pandas
openpyxl
httpx
python-dotenv 